+++ nft-series/src/enumeration.rs
@@ -27,7 +27,7 @@ impl Contract {
     //Query for nft tokens on the contract regardless of the owner using pagination
     pub fn nft_tokens(&self, from_index: Option<u128>, limit: Option<u64>) -> Vec<JsonToken> {
         //where to start pagination - if we have a from_index, we'll use that - otherwise start from 0 index
-        let start = u128::from(from_index.unwrap_or(0));
+        let start = from_index.unwrap_or(0);
 
         //iterate through each token using an iterator
         self.tokens_by_id
@@ -97,7 +97,7 @@ impl Contract {
     // Paginate through all the series on the contract and return the a vector of JsonSeries
     pub fn get_series(&self, from_index: Option<u128>, limit: Option<u64>) -> Vec<JsonSeries> {
         //where to start pagination - if we have a from_index, we'll use that - otherwise start from 0 index
-        let start = u128::from(from_index.unwrap_or(0));
+        let start = from_index.unwrap_or(0);
 
         //iterate through each series using an iterator
         self.series_by_id
@@ -159,7 +159,7 @@ impl Contract {
         };
 
         //where to start pagination - if we have a from_index, we'll use that - otherwise start from 0 index
-        let start = u128::from(from_index.unwrap_or(0));
+        let start = from_index.unwrap_or(0);
 
         //iterate through the tokens
         tokens
